backend/requirements.txt
flask
tensorflow==2.17.1
pillow
flask-cors
numpy